+++ CHANGELOG.md
@@ -9,13 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Moved to use GitHub Actions for label enforcement
-
 ### Fixed
 
 ### Removed
 
 ### Added
+
+## [3.22.0] - 2022-12-13
+
+### Changed
+
+- Moved to use GitHub Actions for label enforcement
+- Add extra flags for Intel Fortran to allow for stricter builds
+  - `-stand f18` to enable Fortran 2018 Standard compliance
+  - `-diag-error 6188` to cause if(integer) to fail
+  - `-diag-error 6192` to cause logical set to integer to fail
+  - `-diag-disable 5268` to suppress warning for long source lines (which our macros often make)
 
 ## [3.21.0] - 2022-11-28